Initial commit - Histories are exported to a Git repo
And a script returns 401/403/200 to be used with nginx’s
auth_request on a Git repo with git-http-backend, so that
a MediaWiki user can authenticate herself/himself with her/
his password to push to a repo.

Remaining part of the writing flow is to create a Git hook
to write into MediaWiki.